// packages/game-engine/src/games/monodeal/types/errors.ts
export type GameErrorCode =
  | "GAME_NOT_IN_PROGRESS"
  | "GAME_ALREADY_COMPLETED"
  | "NOT_YOUR_TURN"
  | "MUST_DRAW_FIRST"
  | "ALREADY_DRAWN_THIS_TURN"
  | "NO_ACTIONS_REMAINING"
  | "CARD_NOT_IN_HAND"
  | "INVALID_CARD_TYPE"
  | "CANNOT_BANK_CARD"
  | "CANNOT_BANK_PROPERTY"
  | "INVALID_PROPERTY_COLOR"
  | "PROPERTY_SET_NOT_FOUND"
  | "CANNOT_ADD_TO_COMPLETED_SET"
  | "INVALID_WILD_COLOR"
  | "CARD_NOT_IN_SET"
  | "INVALID_ACTION_TARGET"
  | "TARGET_PLAYER_NOT_FOUND"
  | "CANNOT_TARGET_SELF"
  | "SET_IS_NOT_COMPLETE"
  | "SET_IS_ALREADY_COMPLETE"
  | "BUILDING_ON_COMPLETE_SET"
  | "TARGET_CARD_NOT_FOUND"
  | "OFFERED_CARD_NOT_FOUND"
  | "HOUSE_REQUIRES_FULL_SET"
  | "HOTEL_REQUIRES_HOUSE"
  | "CANNOT_ADD_BUILDING_TO_SPECIAL_SET"
  | "RENT_COLOR_NOT_OWNED"
  | "RENT_COLOR_MISMATCH"
  | "MUST_RESOLVE_PENDING_ACTION"
  | "NOT_WAITING_FOR_YOUR_REACTION"
  | "NOT_WAITING_FOR_YOUR_PAYMENT"
  | "INSUFFICIENT_PAYMENT"
  | "CANNOT_PAY_FROM_HAND"
  | "HAND_LIMIT_EXCEEDED"
  | "INVALID_DISCARD_COUNT"
  | "WILD_CARD_REQUIRES_EXISTING_PROPERTY"
  | "CANNOT_BREAK_SET_WITH_BUILDINGS"
  | "TIME_EXTENSION_ALREADY_USED"
  | "UNKNOWN_ERROR";

export class GameEngineError extends Error {
  public readonly code: GameErrorCode;

  constructor(code: GameErrorCode, message: string) {
    super(message);
    this.name = "GameEngineError";
    this.code = code;
  }
}